fix(overview): Only load folders tasks for folders that are on screen
