--- shared/src/containers/ProjectTreeTable/context/ProjectTableContext.ts
@@ -1,106 +1,0 @@
-import { createContext, useContext } from 'react'
-import { useLocation, useNavigate, useParams, useSearchParams } from 'react-router-dom'
-import { EntityMap, TableRow } from '../types/table'
-import {
-  FindInheritedValueFromAncestors,
-  GetAncestorsOf,
-  GetInheritedDependents,
-  FindNonInheritedValues,
-} from '../hooks/useFolderRelationships'
-import { RowId } from '../utils/cellUtils'
-
-import { ProjectTableProviderProps } from './ProjectTableProvider'
-import type { SubtasksManagerProps } from '@shared/components'
-import { SoftErrorAction } from '../types'
-
-export type ToggleExpandAll = (rowIds: RowId[], expand?: boolean) => void
-export type ToggleExpands = (rowIds: RowId[], expand?: boolean) => void
-
-export interface ProjectTableContextType {
-  isInitialized: ProjectTableProviderProps['isInitialized']
-  isLoading: ProjectTableProviderProps['isLoading']
-  users: ProjectTableProviderProps['users']
-  // Attributes
-  attribFields: ProjectTableProviderProps['attribFields']
-  attribFieldsScoped: ProjectTableProviderProps['attribFields']
-  error?: string
-  softError?: string // shows a warning banner but doesn't block the table from rendering
-  softErrorAction?: SoftErrorAction
-
-  scopes: ProjectTableProviderProps['scopes']
-
-  // Data
-  tableData: TableRow[]
-  tasksMap: ProjectTableProviderProps['tasksMap']
-  foldersMap: ProjectTableProviderProps['foldersMap']
-  entitiesMap: ProjectTableProviderProps['entitiesMap']
-  fetchNextPage: ProjectTableProviderProps['fetchNextPage']
-  reloadTableData: ProjectTableProviderProps['reloadTableData']
-  getEntityById: (id: string, field?: string) => EntityMap | undefined // if the entity is not found, we explicity search for the field
-  loadingLinksEntityIds?: Set<string>
-
-  // grouping
-  groups: ProjectTableProviderProps['groups']
-  overrideGroupBy?: ProjectTableProviderProps['overrideGroupBy']
-
-  // Filters
-  queryFilters: ProjectTableProviderProps['queryFilters']
-
-  // Hierarchy
-  showHierarchy: ProjectTableProviderProps['showHierarchy']
-  updateShowHierarchy: ProjectTableProviderProps['updateShowHierarchy']
-  hierarchyOptions?: ProjectTableProviderProps['hierarchyOptions']
-  hierarchyActive?: ProjectTableProviderProps['hierarchyActive']
-
-  // Flat folder view
-  isFlatFolderView?: boolean
-
-  // Expanded state
-  expanded: ProjectTableProviderProps['expanded']
-  setExpanded: ProjectTableProviderProps['setExpanded']
-  toggleExpanded: ProjectTableProviderProps['toggleExpanded']
-  updateExpanded: ProjectTableProviderProps['updateExpanded']
-  toggleExpandAll: ToggleExpandAll
-  toggleExpands: ToggleExpands // expand/collapse multiple rows at once
-
-  // Folder Relationships
-  getInheritedDependents: GetInheritedDependents
-  findInheritedValueFromAncestors: FindInheritedValueFromAncestors
-  findNonInheritedValues: FindNonInheritedValues
-  getAncestorsOf: GetAncestorsOf
-
-  // Context menu
-  contextMenuItems: ProjectTableProviderProps['contextMenuItems']
-
-  // Powerpack context
-  powerpack?: ProjectTableProviderProps['powerpack']
-
-  // remote modules
-  modules: ProjectTableProviderProps['modules']
-
-  // SubtasksManager
-  SubtasksManager?: React.ComponentType<SubtasksManagerProps>
-
-  // player
-  playerOpen?: ProjectTableProviderProps['playerOpen']
-  onOpenPlayer?: ProjectTableProviderProps['onOpenPlayer']
-
-  // Views
-  onResetView?: () => void
-
-  // router hooks
-  useParams?: typeof useParams
-  useNavigate?: typeof useNavigate
-  useLocation?: typeof useLocation
-  useSearchParams?: typeof useSearchParams
-}
-
-export const ProjectTableContext = createContext<ProjectTableContextType | undefined>(undefined)
-
-export const useProjectTableContext = () => {
-  const context = useContext(ProjectTableContext)
-  if (!context) {
-    throw new Error('useProjectTableContext must be used within a ProjectTableProvider')
-  }
-  return context
-}
