+++ shared/src/containers/ProjectTreeTable/context/ProjectTableContextInstance.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import { ProjectTableContextType } from './ProjectTableContext'
+
+export const ProjectTableContext = createContext<ProjectTableContextType | undefined>(undefined)
