+++ src/pages/ProjectOverviewPage/context/ProjectOverviewContextInstance.ts
@@ -1,0 +1,6 @@
+import { ProjectOverviewContextType } from '@shared/containers'
+import { createContext } from 'react'
+
+export const ProjectOverviewContext = createContext<ProjectOverviewContextType | undefined>(
+  undefined,
+)
